brute_force was added and approved

+++ src/main/java/com/javarush/Main.java
@@ -1,13 +1,16 @@
 package com.javarush;
 
+import com.javarush.brute_force.BruteForce;
 import com.javarush.cesarcipher.CesarCipher;
 import com.javarush.cli.CLI;
 import com.javarush.io.FileReader;
 import com.javarush.io.FileWriter;
-import com.javarush.moddefiner.ModeLanguageDefiner;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 public class Main {
-
     public static void main(String[] args) {
         CLI cli = new CLI();
         if (args.length == 2 || args.length == 3) {
@@ -16,33 +19,75 @@ public class Main {
             int key = Integer.parseInt(args[2]);
             System.out.println(command);
             System.out.println(filePath);
+            System.out.println(key);
+            switch (command) {
+                case "encrypt":
+                    FileWriter fileWriter = new FileWriter(filePath);
+                    FileReader fileReader = new FileReader(filePath);
+                    CesarCipher cesarCipher = new CesarCipher();
+                    fileWriter.writeFile(cesarCipher.сesarEncryptorEnglish(fileReader.readFile(), key).toString());
 
+                    break;
+
+                case "decrypt":
+                    Path pathFile = Paths.get(filePath);
+                    if (Files.exists(pathFile)) {
+                        FileReader fileReader1 = new FileReader(filePath);
+                        CesarCipher cesarCipher1 = new CesarCipher();
+                        FileWriter fileWriter1 = new FileWriter(filePath);
+                        fileWriter1.writeFile(cesarCipher1.сesarEncryptorEnglish(filePath, key).toString());
+                    } else {
+                        System.out.println("Entered file doesn't exist!");
+                    }
+
+
+                    break;
+                case "brute_force":
+                    FileReader fileReader2 = new FileReader(filePath);
+                    String encryptedText = fileReader2.readFile();
+
+
+                    break;
+
+
+                default:
+                    System.out.println("Your command was entered incorrect!");
+                    System.out.println("Pay attention and try again...");
+            }
         } else {
             cli.InputUser();
-        }
-        switch (cli.getCommand().toLowerCase()) {
-            case "encrypt":
-                FileWriter fileWriter = new FileWriter(cli.getFilePath());
-                FileReader fileReader = new FileReader(cli.getFilePath());
-                CesarCipher cesarCipher = new CesarCipher();
-                fileWriter.writeFile(cesarCipher.CesarEncryptorEnglish(fileReader.readFile(),cli.getKey()).toString());
 
-                break;
+            switch (cli.getCommand().toLowerCase()) {
+                case "encrypt":
+                    FileWriter fileWriter = new FileWriter(cli.getFilePath());
+                    FileReader fileReader = new FileReader(cli.getFilePath());
+                    CesarCipher cesarCipher = new CesarCipher();
+                    fileWriter.writeFile(cesarCipher.сesarEncryptorEnglish(fileReader.readFile(), cli.getKey()).toString());
+                    break;
 
-            case "decrypt":
-                FileReader fileReader1 = new FileReader(cli.getFilePath());
-                ModeLanguageDefiner modeLanguageDefiner1 = new ModeLanguageDefiner();
-                CesarCipher cesarCipher1 = new CesarCipher();
-                FileWriter fileWriter1 = new FileWriter(cli.getFilePath());
-                fileWriter1.writeFile(cesarCipher1.CesarDecryptorEnglish(fileReader1.readFile(), cli.getKey()).toString());
+                case "decrypt":
+                    Path pathFile = Paths.get(cli.getFilePath());
+                    if (Files.exists(pathFile)) {
+                        FileReader fileReader1 = new FileReader(cli.getFilePath());
+                        CesarCipher cesarCipher1 = new CesarCipher();
+                        FileWriter fileWriter1 = new FileWriter(cli.getFilePath());
+                        fileWriter1.writeFile(cesarCipher1.сesarDecryptorEnglish(fileReader1.readFile(), cli.getKey()).toString());
+                    } else {
+                        System.out.println("Entered file doesn't exist!");
+                    }
 
+                    break;
+                case "brute_force":
+                    FileReader fileReader2 = new FileReader(cli.getFilePath());
+                    String encryptedText = fileReader2.readFile();
+                    BruteForce bruteForce = new BruteForce(encryptedText, cli.getFilePath());
+                    bruteForce.brutForce();
 
-                break;
-            case "brute_force":
-
-            default:
-                System.out.println("Your command was entered incorrect!");
-                System.out.println("Pay attention and try again...");
+                    break;
+                default:
+                    System.out.println("Your command was entered incorrect!");
+                    System.out.println("Pay attention and try again...");
+            }
         }
     }
 }
